100% this is the final working copy

+++ position.cpp
@@ -9,8 +9,8 @@ uint64_t zobrist::zobrist_table[NPIECES][NSQUARES];
 //Initializes the zobrist table with random 64-bit numbers
 void zobrist::initialise_zobrist_keys() {
 	PRNG rng(70026072);
-	for (int i = 0; i < NPIECES; i++)
-		for (int j = 0; j < NSQUARES; j++)
+	for (int i = 0; i < (int)NPIECES; i++)
+		for (int j = 0; j < (int)NSQUARES; j++)
 			zobrist::zobrist_table[i][j] = rng.rand<uint64_t>();
 }
